check empty tmp dir before deleting in addTaxon

+++ fdog/addTaxa.py
@@ -154,7 +154,8 @@ def main():
             genome_file = '%s/%s/%s.fa' % (searchpath, name_dict[f], name_dict[f])
             add_taxon_fn.create_annoFile(annopath, genome_file, cpus, force)
         if os.path.exists('%s/tmp' % annopath):
-            shutil.rmtree('%s/tmp' % annopath)
+            if not os.listdir('%s/tmp' % annopath):
+                shutil.rmtree('%s/tmp' % annopath)
         out_msg = '%s, %s' % (out_msg, annopath)
 
     end = time.time()
